Move things into their appropriate packages
