+++ blueocean-pipeline-api/src/main/java/io/jenkins/blueocean/rest/impl/pipeline/credential/BlueOceanDomainSpecification.java
@@ -1,8 +1,7 @@
-package io.jenkins.blueocean.pipeline.credential;
+package io.jenkins.blueocean.rest.impl.pipeline.credential;
 
 import com.cloudbees.plugins.credentials.domains.DomainRequirement;
 import com.cloudbees.plugins.credentials.domains.DomainSpecification;
-import io.jenkins.blueocean.pipeline.credential.BlueOceanDomainRequirement;
 
 import javax.annotation.Nonnull;
